% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{.extract_inventory_design}
\alias{.extract_inventory_design}
\title{Extract a single constant sampling-design value from an inventory table}
\usage{
.extract_inventory_design(data)
}
\description{
Extract a single constant sampling-design value from an inventory table
}
\keyword{internal}
