% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{.extract_inventory_metadata}
\alias{.extract_inventory_metadata}
\title{Extract a single constant metadata value from an inventory table}
\usage{
.extract_inventory_metadata(data, name)
}
\description{
Extract a single constant metadata value from an inventory table
}
\keyword{internal}
