% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_curve.R
\name{precision_curve}
\alias{precision_curve}
\title{Build a precision-versus-sample-size curve}
\usage{
precision_curve(x, variable, n_min = 5L, n_max = NULL)
}
\arguments{
\item{x}{A `forest_adequacy` object.}

\item{variable}{One variable name assessed in `x`.}

\item{n_min}{Minimum sample size shown, default 5.}

\item{n_max}{Maximum sample size. If NULL, chosen automatically.}
}
\value{
A data.frame with sample size and expected relative error percent.
}
\description{
Build a precision-versus-sample-size curve
}
